+++ stories/dropzone.stories.js
@@ -5,7 +5,7 @@ import { Draggable, DragDropContext, Droppable } from "react-beautiful-dnd";
 
 import { FileItem } from "../src/FileItem";
 import FilesField from "../src/FilesField";
-import FormTemplate from "../src/FormTemplate";
+import FormTemplate from "../tests/FormTemplate";
 
 export default {
 	title: "Example using dropzone and react beautiful dnd",
